Music, song text and rest of help options implementation

+++ Assets/Scripts/Game/Line.cs
@@ -8,18 +8,18 @@ namespace Scripts
 {
     public class Line
     {
-        private double startTime;
-        private double endTime;
+        private float startTime;
+        private float endTime;
         private string text;
 
-        public double StartTime { get => startTime; }
-        public double EndTime { get => endTime; }
+        public float StartTime { get => startTime; }
+        public float EndTime { get => endTime; }
         public string Text { get => text; }
 
         public Line(Dictionary<string, string> lineDict)
         {
-            this.startTime = double.Parse(lineDict["start"], System.Globalization.CultureInfo.InvariantCulture);
-            this.endTime = double.Parse(lineDict["end"], System.Globalization.CultureInfo.InvariantCulture);
+            this.startTime = float.Parse(lineDict["start"], System.Globalization.CultureInfo.InvariantCulture);
+            this.endTime = float.Parse(lineDict["end"], System.Globalization.CultureInfo.InvariantCulture);
             this.text = lineDict["text"];
         }
     }
